+++ src/parser/dissectors/application/types.ts
@@ -1,6 +1,7 @@
+import type { FTPStream } from "./FTP/types";
 import type { HTTPStream } from "./HTTP1.1/types";
 
-export type ApplicationLayer = HTTPStream | UnknownApplicationLayer;
+export type ApplicationLayer = HTTPStream | FTPStream | UnknownApplicationLayer;
 
 export type UnknownApplicationLayer = {
   type: "unknown-application",